+++ src/util.c
@@ -208,7 +208,7 @@ static void paginate_msg(const char *fmt, va_list ap,
    const char *p = filtered;
    int col = start;
    while (*p != '\0') {
-      if ((*p == '\n') || (isspace((uint8_t)*p) && col >= right)) {
+      if ((*p == '\n') || (isspace((int)*p) && col >= right)) {
          // Can break line here
          fputc('\n', stderr);
          for (col = 0; col < left; col++)
